requirements.txt
flask
flask-cors
gunicorn
requests
pyjwt
pyjwt[crypto]
python-dotenv
sqlite3
bcrypt
